requirements.txt
-e git+https://github.com/django-nonrel/django-nonrel.git@c68fd7ecc9882917d50c44e3bf1e07a8f183d7d5#egg=Django-dev
PIL==1.1.7
argparse==1.2.1
distribute==0.6.24
django-annoying==0.7.6
-e git+https://github.com/FTW/django-mongodb-engine.git@52257b5d90dcfaa564e708264a9cfe591d301fe6#egg=django_mongodb_engine-dev
-e git+https://github.com/MirkoRossini/django-redis-engine.git@fc6fddbcf080b0652efb728314e67b21d4b8186a#egg=django_redis_engine-dev
-e git+https://github.com/danpetrikin/django-singletons@f35aad54eb42c5c70f58a0b7e05df28b9635ceb7#egg=django_singletons-dev
-e git+https://github.com/mburstiner/djangotoolbox@73ffcb76ee606bb52505124f88454036fa938daa#egg=djangotoolbox-dev
gunicorn==0.15.0
meld3==0.6.9
pymongo==2.3
redis==2.7.1
requests==0.14.2
stripe==1.7.5
supervisor==3.0b1
wsgiref==0.1.2
